requirements.txt
aiogram==3.24.0
Pillow==12.1.0